REM conversion part 1
REM conversion with px as document root font-size, which is not recommened because it takes away the feature for browsers & screen readers to have a specific font-size. Instead of them being able to make it larger or smaller the font size will always be our docuemtn root font size instead.