ResponsiveRem In practice
The root font size for font & length with 75% would be 12px. 100% is browser default font-size which is 16px. You do %/100 / n/16 to get font-size in pixle conversion. Any rem is divided by font-size because it is root font size so 10/30px. With this we can increase or decrease the size by only tweaking one value, the root-font-size. How convient!